moved common dir out of bankapi
